Bugfix: data loss when target contains trailing slash
I pretty comprehensively tested this with a mix of leading a trailing slashes and found that when the symlink already exists, if the target in the composer json dictionary has a trailing slash, it gets deleted (MacOS anyway).

I can't think of any instance (file or folder) where the destination would need a trailing slash, so I'm removing it.